WIP CRUD trait; user account and directory modules using CRUD trai
